test(storybook): modularize visual-test runtime
Closes #6763
