+++ packages/react/storybook/visual-test/constants.ts
@@ -1,0 +1,9 @@
+export const VISUAL_TEST_FIXED_NOW = Date.UTC(2025, 3, 15, 13, 24, 36)
+export const VISUAL_TEST_LOCALE = "en-US"
+export const VISUAL_TEST_ONLINE = true
+export const VISUAL_TEST_RANDOM_SEED = 6763
+export const VISUAL_TEST_TIME_ZONE = "UTC"
+export const VISUAL_TEST_USER_AGENT =
+  "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (like Gecko) Chrome/124.0.0.0 Safari/537.36"
+export const VISUAL_TEST_VIEWPORT_WIDTH = 1200
+export const VISUAL_TEST_UUID_PREFIX = "00000000-0000-4000-8000-"
